ragcore/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29
python-dotenv>=1.0
pydantic>=2.6
pydantic-settings>=2.2
sentence-transformers>=2.6
qdrant-client>=1.8
rank-bm25>=0.2.2
FlashRank>=0.2
google-generativeai>=0.5
gradio>=4.20
pypdf>=4.1
beautifulsoup4>=4.12
httpx>=0.27
python-multipart>=0.0.9
python-dateutil>=2.9
ruff>=0.3
pytest>=8.0
